ai-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.4
python-dotenv==1.1.1
psycopg[binary]>=3.2.0
